extract new status description into separate spec
