+++ spec/models/work_package_spec.rb
@@ -979,130 +979,6 @@ describe WorkPackage, type: :model do
     end
   end
 
-  describe '#new_statuses_allowed_to' do
-    let(:role) { FactoryGirl.create(:role) }
-    let(:type) { FactoryGirl.create(:type) }
-    let(:user) { FactoryGirl.create(:user) }
-    let(:other_user) { FactoryGirl.create(:user) }
-    let(:statuses) { (1..5).map { |_i| FactoryGirl.create(:status) } }
-    let(:priority) { FactoryGirl.create :priority, is_default: true }
-    let(:status) { statuses[0] }
-    let(:project) do
-      FactoryGirl.create(:project, types: [type]).tap { |p| p.add_member(user, role).save }
-    end
-    let(:workflow_a) {
-      FactoryGirl.create(:workflow, role_id: role.id,
-                                    type_id: type.id,
-                                    old_status_id: statuses[0].id,
-                                    new_status_id: statuses[1].id,
-                                    author: false,
-                                    assignee: false)
-    }
-    let(:workflow_b) {
-      FactoryGirl.create(:workflow, role_id: role.id,
-                                    type_id: type.id,
-                                    old_status_id: statuses[0].id,
-                                    new_status_id: statuses[2].id,
-                                    author: true,
-                                    assignee: false)
-    }
-    let(:workflow_c) {
-      FactoryGirl.create(:workflow, role_id: role.id,
-                                    type_id: type.id,
-                                    old_status_id: statuses[0].id,
-                                    new_status_id: statuses[3].id,
-                                    author: false,
-                                    assignee: true)
-    }
-    let(:workflow_d) {
-      FactoryGirl.create(:workflow, role_id: role.id,
-                                    type_id: type.id,
-                                    old_status_id: statuses[0].id,
-                                    new_status_id: statuses[4].id,
-                                    author: true,
-                                    assignee: true)
-    }
-    let(:workflows) { [workflow_a, workflow_b, workflow_c, workflow_d] }
-
-    it 'should respect workflows w/o author and w/o assignee' do
-      workflows
-      expect(status.new_statuses_allowed_to([role], type, false, false))
-        .to match_array([statuses[1]])
-      expect(status.find_new_statuses_allowed_to([role], type, false, false))
-        .to match_array([statuses[1]])
-    end
-
-    it 'should respect workflows w/ author and w/o assignee' do
-      workflows
-      expect(status.new_statuses_allowed_to([role], type, true, false))
-        .to match_array([statuses[1], statuses[2]])
-      expect(status.find_new_statuses_allowed_to([role], type, true, false))
-        .to match_array([statuses[1], statuses[2]])
-    end
-
-    it 'should respect workflows w/o author and w/ assignee' do
-      workflows
-      expect(status.new_statuses_allowed_to([role], type, false, true))
-        .to match_array([statuses[1], statuses[3]])
-      expect(status.find_new_statuses_allowed_to([role], type, false, true))
-        .to match_array([statuses[1], statuses[3]])
-    end
-
-    it 'should respect workflows w/ author and w/ assignee' do
-      workflows
-      expect(status.new_statuses_allowed_to([role], type, true, true))
-        .to match_array([statuses[1], statuses[2], statuses[3], statuses[4]])
-      expect(status.find_new_statuses_allowed_to([role], type, true, true))
-        .to match_array([statuses[1], statuses[2], statuses[3], statuses[4]])
-    end
-
-    it 'should respect workflows w/o author and w/o assignee on work packages' do
-      workflows
-      work_package = WorkPackage.create(type_id: type.id,
-                                        status: status,
-                                        priority: priority,
-                                        project: project)
-      expect(work_package.new_statuses_allowed_to(user)).to match_array([statuses[0], statuses[1]])
-    end
-
-    it 'should respect workflows w/ author and w/o assignee on work packages' do
-      workflows
-      work_package = WorkPackage.create(type_id: type.id,
-                                        status: status,
-                                        priority: priority,
-                                        project: project,
-                                        author: user)
-      expect(work_package.new_statuses_allowed_to(user))
-        .to match_array([statuses[0], statuses[1], statuses[2]])
-    end
-
-    it 'should respect workflows w/o author and w/ assignee on work packages' do
-      workflows
-      work_package = WorkPackage.create(type_id: type.id,
-                                        status: status,
-                                        subject: 'test',
-                                        priority: priority,
-                                        project: project,
-                                        assigned_to: user,
-                                        author: other_user)
-      expect(work_package.new_statuses_allowed_to(user))
-        .to match_array([statuses[0], statuses[1], statuses[3]])
-    end
-
-    it 'should respect workflows w/ author and w/ assignee on work packages' do
-      workflows
-      work_package = WorkPackage.create(type_id: type.id,
-                                        status: status,
-                                        subject: 'test',
-                                        priority: priority,
-                                        project: project,
-                                        author: user,
-                                        assigned_to: user)
-      expect(work_package.new_statuses_allowed_to(user))
-        .to match_array([statuses[0], statuses[1], statuses[2], statuses[3], statuses[4]])
-    end
-  end
-
   describe '#add_time_entry' do
     it 'should return a new time entry' do
       expect(stub_work_package.add_time_entry).to be_a TimeEntry
